Make Wi-Fi work
The firmware in `linux-firmware` has only been recently updated to
support the Lenovo Yoga Slim 7x, and it hasn't made it into a release
yet, but we can update to the relevant commit directly.

The `ath12k` module needs to be removed from the initrd because
otherwise it tries to load the firmware too early, and fails because it
only becomes available once the root filesystem is mounted.