Make battery indicator work
Below I describe my current understanding of how this works, though I
could be wrong in some places and there are still a lot of unknowns.

The battery status is requested from the aDSP coprocessor, which is
managed by the qcom_q6v5_pas module, so we have to enable it. We can
only do this though when booting from NVMe, since enabling the aDSP
messes with USB booting. (It is also somehow involved with managing
USB.)

A userspace pd-mapper (Protection Domain Mapper) daemon that
communicates with the aDSP is also needed for configuring it before we
can request the battery status. (A kernel version of pd-mapper also
exists but is currently broken, hopefully though it will be fixed soon
and we can switch over to it.) This commit packages pd-mapper and adds
it as a systemd service.

Some further firmware files are needed by pd-mapper. These firmware
files are just JSON files, and I copied them from Windows. (They are
CRLF and git complains about them, but I did not modify them for now.)
Note that they cannot be compressed using zstd, because pd-mapper does
not support it.

The ISO should be reproducible and have the following SHA256 hash:
a5cdb2c3d328bd53c6e2e509aad77f8a403d8e2dd807cc9c00c446e2db37c313