Enable mdformat to format markdown files
You might notice that now the lists have all items numbered as "1". It
turns out that this is actually a feature and there is an option to
disable it. I think it is meant to reduce diff noise when
adding/removing items from lists... which I guess makes sense? Just
takes a bit of getting used to.