Add camera support
This new post-fetch hook is definitely a bit... interesting. First this
whole thing of exporting a variable to be evaluated after fetching in
the preFetch hook is weird, but this is what a comment in nixpkgs
suggest.

The fetching does take a couple minutes now so it's definitely a bit
slower than before. It's not a deal breaker compared to the 20+ mins it
takes to compile the kernel, but optimizations would be welcome of
course.

The `--depth` numbers on the fetch commands look a bit arbitrary, they
are a poor imitation of the `--shallow-exclude` flag that I think should
be used here, but for some reason none of the git remotes I tried worked
with that flag.

The fetching from the two remotes could also in theory be done in
parallel, but I did not find a way to specify multiple remotes with
multiple refs and different `--depth` flags to `git fetch`.

As for the camera patches, it's obviously from a debug branch with
various extra debug prints and stuff added. We could filter those if we
really wanted, but I think with such an experimental driver the debug
messages can be helpful for users too. The branch also touches
johan_defconfig but I think that's fine for now. I took a quick look at
the commits and did not see anything that would be definitely
undesirable.

We also cherry-pick the `platform-qcom-media-for-6.16` tag that was
posted on the mailing list to be pulled into 6.16, since the camera
patches depend on it. This (together with some extra patches in the
camera branch) should also in theory bring video decoding acceleration
support, but I think that still needs some firmware to work, and I
haven't looked into it yet.